% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_helpers.R
\name{fill_na}
\alias{fill_na}
\title{Locate flow gauges on a channel network}
\usage{
fill_na(rst, ctch, layer_name = names(rst))
}
\arguments{
\item{rst}{SpatRast (or path to file) to infill}

\item{ctch}{SpatRast (or path to file) defining non-NA cells}

\item{layer_name}{name of layer to infill}
}
\description{
Locate flow gauges on a channel network
}
\details{
\code{rst} and \code{ctch} should have the same extent and resolution. The \code{NA} values in rst
are infilled using the most common neighbouring value throuhg multiple passes of focal.
}
