% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_channel.R
\name{locate_gauges}
\alias{locate_gauges}
\title{Locate flow gauges on a channel network}
\usage{
locate_gauges(chn, gauges, gauge_name = "name", max_dist = 100)
}
\arguments{
\item{chn}{the channel object}

\item{gauges}{either points of gauge locations of polygons of catchment boundaries}

\item{gauge_name}{feild name in gauges to use for name of gauge}

\item{max_dist}{maxium distance between gauge points and channel}
}
\description{
Locate flow gauges on a channel network
}
\details{
If gauges are points the nearest channel length within max_dist is chosen. If catchment polygons are given all channel lengths leaving the catchment are returned.
}
