% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_channel.R
\name{merge_channels}
\alias{merge_channels}
\title{Merge two channel objects}
\usage{
merge_channels(x, y, outlets = NULL, verbose = FALSE)
}
\arguments{
\item{x}{the channel object to burn into}

\item{y}{the channel object to be burnt in}

\item{outlets}{names of the channel lengths that are outlets, must be in x}

\item{verbose}{show progress}
}
\description{
Merge on channel object into another by burning them in
}
\details{
Elements in x are either cropped, or fully removed if they lie under y. Attempts are made to ensure that the start and end Nodes are suitably replaced
}
