% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_channel.R
\name{simplify_channel}
\alias{simplify_channel}
\title{simplify and channel object}
\usage{
simplify_channel(
  chn,
  simplify_length = 100,
  outlets = NULL,
  strict_routing = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{chn}{the channel object}

\item{simplify_length}{minimum length of channel sections to remove}

\item{outlets}{specify outlet reaches to check the simplified channel on exit}

\item{strict_routing}{see details}

\item{verbose}{print a higher level of output}
}
\description{
Simplify a channel object by merging spatial objects
}
\details{
strict_routing will maintain the flow pathways by only removing nodes with a single input and output, otherwise the routing may alter the effective flow lengths.
The is_wb feild in the data is used to flag objects that shouldn't be merged..
}
