% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_channel.R
\name{trim_channel}
\alias{trim_channel}
\title{Trim channel objects}
\usage{
trim_channel(chn, outlets, removed = FALSE)
}
\arguments{
\item{chn}{the channel object}

\item{outlets}{names of the outlets to consider}

\item{removed}{should the removed objects be returned}
}
\description{
Trim channel objects to those draining to named outlets
}
\details{
Starting with the named outlet objects the channel network is passed up using the
connectivity given by the start and end node, only those flaged as connected are returned.
}
