% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_channel.R
\name{convert_channel}
\alias{convert_channel}
\title{Function for assisting in the conversion of object to be suitable channel inputs to a dynatopGIS object}
\usage{
convert_channel(
  chn,
  property_names = c(name = "DRN_ID", length = "length", startNode = "startNode", endNode
    = "endNode", width = "width", slope = "slope"),
  default_width = 2,
  default_slope = 0.001,
  simplify_length = 0,
  wb = NULL
)
}
\arguments{
\item{chn}{a SpatVect object or a file which can read by terra::vect to create one}

\item{property_names}{a named vector of containing the columns of existing data properties required in the final SpatialPolygonsDataFrame}

\item{default_width}{the width in m to be used for buffering lines to produce polygons}

\item{default_slope}{the slope in m/m to be used when none is provided}
}
\description{
Converts SpatVect of the channel network into SpatVect containing polgons suitable for dynatopGIS, optionally adding in water bodies and simplifying.
}
\details{
If the property_names vector contains a width this is used for buffering lines to produce polygons, otherwise the default_width value is used.
}
\examples{
channel_file <- system.file("extdata", "SwindaleRiverNetwork.shp",
package="dynatopGIS", mustWork = TRUE)
vect_lines <- terra::vect(channel_file)
property_names <- c(name="identifier",endNode="endNode",startNode="startNode",length="length")
chn <- convert_channel(vect_lines,property_names)
}
